Time map view development:
More sophisticated handling of theme variables. Number of breadcrumbs
on My Map Track is now defined by number of theme tags defined in the
variable. Time map view is now supported by record types 51 (Historical
Event), 165 (KML file) and 103 (KML map). For type 103 pointers to 165
will be rendered on the time map, as well as related records of type 51.
Also, handling for accurate rendering of the legend for related items as
well as map track is introduced.